Add more authorization checks for admin/groups and admin/collections

+++ app/controllers/admin/collections_controller.rb
@@ -26,6 +26,7 @@ module Admin
     end
 
     def destroy
+      authorize(collection)
       flash[:notice] = %Q{Collection "#{collection.name}" has been deleted}
 
       collection.destroy
